Move `BoundedContext.newName()` to `BoundedContextNames`.
