+++ server/src/test/java/io/spine/server/storage/memory/InMemoryStorageFactoryTest.java
@@ -24,7 +24,7 @@ import io.spine.core.BoundedContextName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static io.spine.server.BoundedContext.newName;
+import static io.spine.core.BoundedContextNames.newName;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
